chore(grunt): use rimraf to remove node_modules on Windows

+++ scripts/npm/install-dependencies.sh
@@ -9,7 +9,10 @@ if diff -q $SHRINKWRAP_FILE $SHRINKWRAP_CACHED_FILE; then
   echo 'No shrinkwrap changes detected. npm install will be skipped...';
 else
   echo 'Blowing away node_modules and reinstalling npm dependencies...'
-  rm -rf node_modules
+  if !(node 'scripts/npm/call-rimraf.js'); then
+    rm -rf node_modules;
+    echo 'cleaned node_modules using rm -rf';
+  fi
   npm install
   cp $SHRINKWRAP_FILE $SHRINKWRAP_CACHED_FILE
   echo 'npm install successful!'
